+++ core/models.py
@@ -97,5 +97,9 @@ class Housing(BaseModel):
     rate_pct: float = 6.75
     term_years: int = 30
     tax_rate_pct: float = 1.0
+    hoi_rate_pct: float | None = 0.0
     hoi_annual: float = 1200.0
     hoa_monthly: float = 0.0
+    finance_upfront: bool = True
+    credit_score: float = 760.0
+    first_use_va: bool = True
