+++ core/version.py
@@ -1,0 +1,6 @@
+from importlib import metadata
+
+try:
+    __version__ = metadata.version("amalo")
+except metadata.PackageNotFoundError:  # pragma: no cover - during local dev
+    __version__ = "0.1.0"
