+++ tests/test_payment_ui.py
@@ -3,7 +3,6 @@ from core.calculators import monthly_payment
 
 
 def housing_app():
-    import streamlit as st
     import app
 
     app.render_property_column()
